// audtech-backend/src/common/enums/resultado-item.enum.ts
export enum ResultadoItem {
  CONFORME = 'CONFORME',
  NAO_CONFORME = 'NAO_CONFORME',
  NAO_APLICAVEL = 'NAO_APLICAVEL',
}

export enum CategoriaPendencia {
  DOCUMENTAL = 'DOCUMENTAL',
  ESTRUTURAL = 'ESTRUTURAL',
  AMBIENTAL = 'AMBIENTAL',
}

export enum TipoArquivo {
  JPG = 'JPG',
  PNG = 'PNG',
  PDF = 'PDF',
}